refactoring

+++ Menu/Models/DAL/Domain/Article.cs
@@ -1,6 +1,8 @@
 ﻿
 
 
+using System.Collections.Generic;
+
 namespace Menu.Models.DAL.Domain
 {
     public class Article
@@ -10,11 +12,12 @@ namespace Menu.Models.DAL.Domain
         public string Body { get; set; }
         public string MainImagePath { get; set; }
 
+        //public List<long> AdditionalImagesPath { get; set; }
 
         public bool Followed { get; set; }
 
 
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public User User { get; set; }
 
     }
